backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.9.0
celery==5.5.2
cffi==2.0.0
channels==4.2.0
channels_redis==4.2.1
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.7
daphne==4.2.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.4.0
graphql-core==3.2.8
hyperlink==21.0.0
idna==3.13
Incremental==24.11.0
kombu==5.5.4
msgpack==1.1.2
packaging==26.1
pillow==11.2.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
pyOpenSSL==26.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
qrcode==8.1
redis==7.4.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
strawberry-graphql==0.262.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2026.1
ujson==5.12.0
vine==5.1.0
wcwidth==0.6.0
zope.interface==8.3
